+++ src/exporters/prometheus/httpd.go
@@ -129,8 +129,8 @@ func (e *Prometheus) ServeMetrics(w http.ResponseWriter, r *http.Request) {
 	}
 
 	duration := time.Since(start)
-	e.Metadata.SetValueSS("time", "render", float32(duration.Seconds()))
-	e.Metadata.SetValueSS("count", "render", float32(count))
+	e.Metadata.SetValueSS("time", "render", float64(duration.Seconds()))
+	e.Metadata.SetValueSS("count", "render", float64(count))
 
 	md, _ := e.Render(e.Metadata)
 	data = append(data, md...)
